+++ volmdlr/surfaces.py
@@ -25,7 +25,7 @@ import volmdlr.utils.common_operations as vm_common_operations
 import volmdlr.utils.intersections as vm_utils_intersections
 import volmdlr.utils.parametric as vm_parametric
 from volmdlr import display, edges, grid, wires, curves
-from volmdlr.core import EdgeStyle, point_in_list
+from volmdlr.core import EdgeStyle
 from volmdlr.nurbs.core import evaluate_surface, derivatives_surface, point_inversion, find_multiplicity
 from volmdlr.nurbs.fitting import approximate_surface, interpolate_surface
 from volmdlr.nurbs.operations import split_surface_u, split_surface_v
@@ -1108,8 +1108,7 @@ class Surface3D(DessiaObject):
         for primitive in contour3d.primitives:
             primitive_plane_intersections = self.edge_intersections(primitive)
             for primitive_plane_intersection in primitive_plane_intersections:
-                if not volmdlr.core.point_in_list(primitive_plane_intersection,
-                                                  outer_contour_intersections_with_plane):
+                if not primitive_plane_intersection.in_list(outer_contour_intersections_with_plane):
                     outer_contour_intersections_with_plane.append(primitive_plane_intersection)
         return outer_contour_intersections_with_plane
 
@@ -1135,7 +1134,7 @@ class Surface3D(DessiaObject):
         method_name = f'{self.__class__.__name__.lower()[:-2]}_intersections'
         if hasattr(other_surface, method_name):
             return getattr(other_surface, method_name)(self)
-        raise (f'No method available for calculating intersections between {self.__class__} and '
+        raise NotImplementedError (f'No method available for calculating intersections between {self.__class__} and '
                f'{other_surface.__class__}')
 
     def line_intersections(self, line: curves.Line3D):
@@ -1197,7 +1196,7 @@ class Surface3D(DessiaObject):
         for curve_plane_intersection in curve_plane_intersections:
             inters = curve_plane_intersection.curve_intersections(curve)
             for intersection in inters:
-                if not volmdlr.core.point_in_list(intersection, intersections):
+                if not intersection.in_list(intersections):
                     intersections.append(intersection)
         return intersections
 
@@ -2548,7 +2547,7 @@ class CylindricalSurface3D(PeriodicalSurface):
                     all_generatrices_intersecting = False
                 for index, point in enumerate(linseg_intersections):
                     if other_surface.point_distance(point) < 1e-6 and \
-                            not volmdlr.core.point_in_list(point, lists_intersections[index]):
+                            not point.in_list(lists_intersections[index]):
                         lists_intersections[index].append(point)
             return lists_intersections, all_generatrices_intersecting
 
@@ -2565,13 +2564,13 @@ class CylindricalSurface3D(PeriodicalSurface):
                 for point in (
                         cylinder_generatrices_point_intersections[0] + cylinder_generatrices_point_intersections[1] +
                         cone_generatrices_point_intersections[0] + cone_generatrices_point_intersections[1]):
-                    if not volmdlr.core.point_in_list(point, intersections_points[0]):
+                    if not point.in_list(intersections_points[0]):
                         intersections_points[0].append(point)
         elif not all_cone_generatrices_intersecting_cylinder:
             intersections_points = [[]]
             for point in (cylinder_generatrices_point_intersections[0] + cylinder_generatrices_point_intersections[1] +
                           cone_generatrices_point_intersections[0] + cone_generatrices_point_intersections[1]):
-                if not volmdlr.core.point_in_list(point, intersections_points[0]):
+                if not point.in_list(intersections_points[0]):
                     intersections_points[0].append(point)
         list_curves = []
         for list_points in intersections_points:
@@ -2610,6 +2609,82 @@ class CylindricalSurface3D(PeriodicalSurface):
         if math.isclose(new_point.x ** 2 + new_point.y ** 2, self.radius ** 2, abs_tol=abs_tol):
             return True
         return False
+
+    def _sphere_cylinder_tangent_intersections(self, frame, distance_axis_sphere_center):
+        """
+        Gets the intersections between a sphere tangent to the cylinder.
+
+        :param frame: frame for local calculations. Frame is such that w is the cylinder axis,
+        and u passes through the sphere's center.
+        :param distance_axis_sphere_center: distance of sphere's center to cylinder axis.
+        :return: return a list with the intersecting curves.
+        """
+        curves_ = []
+        for phi_range in [(0, math.pi), (math.pi, 2*math.pi), (2*math.pi, 3*math.pi), (3*math.pi, 4*math.pi)]:
+            phi = npy.linspace(phi_range[0], phi_range[1], 100)
+            intersection_points = [volmdlr.Point3D(x_comp, y_comp, z_comp)
+                                   for x_comp, y_comp, z_comp in zip(
+                                       self.radius * npy.cos(phi), self.radius * npy.sin(phi),
+                                       2 * math.sqrt(distance_axis_sphere_center*self.radius)*npy.cos(phi / 2))]
+            bspline = edges.BSplineCurve3D.from_points_interpolation(intersection_points, 4, centripetal=False)
+            curves_.append(bspline)
+        global_intersections = [edge.frame_mapping(frame, 'old') for edge in curves_]
+        return global_intersections
+
+    def sphericalsurface_intersections(self, spherical_surface: 'SphericalSurface3D'):
+        """
+        Cylinder Surface intersections with a Spherical surface.
+
+        :param spherical_surface: intersecting sphere.
+        :return: list of intersecting curves.
+        """
+        curves_ = []
+        line_axis = curves.Line3D.from_point_and_vector(self.frame.origin, self.frame.w)
+        distance_axis_sphere_center = line_axis.point_distance(spherical_surface.frame.origin)
+        if distance_axis_sphere_center < self.radius:
+            if distance_axis_sphere_center + spherical_surface.radius < self.radius:
+                return []
+            if math.isclose(distance_axis_sphere_center, 0.0, abs_tol=1e-6):
+                if math.isclose(self.radius, spherical_surface.radius):
+                    return [spherical_surface.get_circle_at_z(0)]
+                z_plane_position = math.sqrt(spherical_surface.radius**2 - self.radius**2)
+                circle1 = spherical_surface.get_circle_at_z(z_plane_position)
+                circle2 = spherical_surface.get_circle_at_z(-z_plane_position)
+                return [circle1, circle2]
+        if distance_axis_sphere_center - spherical_surface.radius > self.radius:
+            return []
+        point_projection, _ = line_axis.point_projection(spherical_surface.frame.origin)
+        vector = (spherical_surface.frame.origin - point_projection).unit_vector()
+        frame = volmdlr.Frame3D(point_projection, vector, self.frame.w.cross(vector), self.frame.w)
+        if math.isclose(distance_axis_sphere_center + self.radius, spherical_surface.radius, abs_tol=1e-6):
+            return self._sphere_cylinder_tangent_intersections(frame, distance_axis_sphere_center)
+        b = (spherical_surface.radius**2 - self.radius**2 -
+             distance_axis_sphere_center**2) / (2*distance_axis_sphere_center)
+        if spherical_surface.radius > self.radius + distance_axis_sphere_center:
+            phi_0 = 0
+            phi_1 = 2*math.pi
+            two_curves = True
+        else:
+            phi_0 = math.acos(-b/self.radius)
+            phi_1 = phi_0-0.000001
+            phi_0 = -phi_0+0.000001
+            two_curves = False
+        phi = npy.linspace(phi_0, phi_1, 400)
+        x_components = self.radius*npy.cos(phi)
+        y_components = self.radius*npy.sin(phi)
+        z_components1 = npy.sqrt(2*distance_axis_sphere_center*(b + x_components))
+        inters_points1 = [volmdlr.Point3D(x_comp, y_comp, z_comp)
+                          for x_comp, y_comp, z_comp in zip(x_components, y_components, z_components1)]
+        inters_points2 = [volmdlr.Point3D(x_comp, y_comp, -z_comp)
+                          for x_comp, y_comp, z_comp in zip(x_components, y_components, z_components1)]
+        inters_points = [inters_points1, inters_points2]
+        if not two_curves:
+            inters_points = vm_common_operations.separate_points_by_closeness(inters_points[0]+inters_points[1])
+        for list_points in inters_points:
+            bspline = edges.BSplineCurve3D.from_points_interpolation(list_points, 4, centripetal=False)
+            curves_.append(bspline)
+        global_intersections = [edge.frame_mapping(frame, 'old') for edge in curves_]
+        return global_intersections
 
 
 class ToroidalSurface3D(PeriodicalSurface):
@@ -3295,14 +3370,14 @@ class ToroidalSurface3D(PeriodicalSurface):
         for arc in arcs:
             intersections = cylindrical_surface.circle_intersections(arc)
             for intersection in intersections:
-                if not volmdlr.core.point_in_list(intersection, points_intersections):
+                if not intersection.in_list(points_intersections):
                     points_intersections.append(intersection)
         for edge in cylindrical_surface.get_generatrices(self.outer_radius * 3, 300):
             # \
             #    cylindrical_surface.get_circle_generatrices(72, self.outer_radius * 3):
             intersections = self.edge_intersections(edge)
             for point in intersections:
-                if not volmdlr.core.point_in_list(point, points_intersections):
+                if not point.in_list(points_intersections):
                     points_intersections.append(point)
         return points_intersections
 
@@ -3376,7 +3451,7 @@ class ToroidalSurface3D(PeriodicalSurface):
         for edge in conical_surface.get_generatrices(self.outer_radius * 3, 300):
             intersections = self.edge_intersections(edge)
             for point in intersections:
-                if not volmdlr.core.point_in_list(point, points_intersections):
+                if not point.in_list(points_intersections):
                     points_intersections.append(point)
         return points_intersections
 
@@ -3413,11 +3488,6 @@ class ToroidalSurface3D(PeriodicalSurface):
         for arc in arcs:
             intersections = spherical_surface.circle_intersections(arc)
             intersection_points.extend(intersections)
-        # for edge in spherical_surface._circle_generatrices(50):
-        #     intersections = self.circle_intersections(edge)
-        #     for point in intersections:
-        #         if not volmdlr.core.point_in_list(point, intersection_points):
-        #             intersection_points.append(point)
         return intersection_points
 
     def sphericalsurface_intersections(self, spherical_surface: 'SphericalSurface3D'):
@@ -3440,7 +3510,7 @@ class ToroidalSurface3D(PeriodicalSurface):
             curves_.append(bspline)
         return curves_
 
-      
+
 class ConicalSurface3D(PeriodicalSurface):
     """
     The local plane is defined by (theta, z).
@@ -3482,6 +3552,15 @@ class ConicalSurface3D(PeriodicalSurface):
             list_generatrices.append(wire)
         return list_generatrices
 
+    def get_circle_generatrices(self, z, number_circles: int):
+        circles = []
+        for i_z in npy.linspace(0, z, number_circles):
+            i_frame = self.frame.translation(volmdlr.Vector3D(0, 0, i_z))
+            radius = i_z * math.tan(self.semi_angle)
+            circle = curves.Circle3D(i_frame, radius)
+            circles.append(circle)
+        return circles
+
     def plot(self, ax=None, edge_style: EdgeStyle = EdgeStyle(color='grey', alpha=0.5), **kwargs):
         """
         Plots the ConicalSurface3D.
@@ -3491,16 +3570,10 @@ class ConicalSurface3D(PeriodicalSurface):
             fig = plt.figure()
             ax = fig.add_subplot(111, projection='3d')
 
-        self.frame.plot(ax=ax, ratio=z)
-        x = z * math.tan(self.semi_angle)
-        # point1 = self.frame.local_to_global_coordinates(volmdlr.Point3D(-x, 0, -z))
-        point1 = self.frame.origin
-        point2 = self.frame.local_to_global_coordinates(volmdlr.Point3D(x, 0, z))
-        generatrix = edges.LineSegment3D(point1, point2)
-        for i in range(37):
-            theta = i / 36. * volmdlr.TWO_PI
-            wire = generatrix.rotation(self.frame.origin, self.frame.w, theta)
-            wire.plot(ax=ax, edge_style=edge_style)
+        line_generatrices = self.get_generatrices(z, 36)
+        circle_generatrices = self.get_circle_generatrices(z, 50)
+        for edge in line_generatrices + circle_generatrices:
+            edge.plot(ax, edge_style)
         return ax
 
     @classmethod
@@ -3927,6 +4000,43 @@ class ConicalSurface3D(PeriodicalSurface):
         """
         return [curves.Line2D(volmdlr.Point2D(-math.pi, 0), volmdlr.Point2D(math.pi, 0))]
 
+    def _spherical_intersection_points(self, spherical_surface: 'SphericalSurface3D'):
+        """
+        Gets the points of intersections between the spherical surface and the toroidal surface.
+
+        :param spherical_surface: other Conical Surface 3d.
+        :return: points of intersections.
+        """
+        cyl_generatrices = self.get_generatrices(spherical_surface.radius*4, 200) +\
+                           self.get_circle_generatrices(200, spherical_surface.radius*4)
+        intersection_points = []
+        for gene in cyl_generatrices:
+            intersections = spherical_surface.edge_intersections(gene)
+            for intersection in intersections:
+                if not intersection.in_list(intersection_points):
+                    intersection_points.append(intersection)
+        return intersection_points
+
+    def sphericalsurface_intersections(self, spherical_surface: 'SphericalSurface3D'):
+        """
+        Cylinder Surface intersections with a Spherical surface.
+
+        :param spherical_surface: intersecting sphere.
+        :return: list of intersecting curves.
+        """
+        intersection_points = self._spherical_intersection_points(spherical_surface)
+        if not intersection_points:
+            return []
+        inters_points = vm_common_operations.separate_points_by_closeness(intersection_points)
+        curves_ = []
+        for list_points in inters_points:
+            bspline = edges.BSplineCurve3D.from_points_interpolation(list_points, 4, centripetal=False)
+            if isinstance(bspline.simplify, edges.FullArc3D):
+                curves_.append(bspline.simplify)
+                continue
+            curves_.append(bspline)
+        return curves_
+
 
 class SphericalSurface3D(PeriodicalSurface):
     """
@@ -3964,6 +4074,24 @@ class SphericalSurface3D(PeriodicalSurface):
             circles.append(circle)
         return circles
 
+    def _circle_generatrices_xy(self, number_circles: int):
+        """
+        Gets the sphere circle generatrices in the a parallel planes.
+
+        :param number_circles: number of circles to be created.
+        :return: List of Circle 3D.
+        """
+        circles = []
+        initial_center = self.frame.origin.translation(-self.frame.w*self.radius)
+        for i in npy.linspace(0, 2 * self.radius, number_circles):
+            center = initial_center.translation(self.frame.w * i)
+            frame = volmdlr.Frame3D(center, self.frame.u, self.frame.v, self.frame.w)
+            dist = center.point_distance(self.frame.origin)
+            circle_radius = math.sqrt(self.radius ** 2 - dist ** 2)
+            circle = curves.Circle3D(frame, circle_radius)
+            circles.append(circle)
+        return circles
+
     @property
     def domain(self):
         """Returns u and v bounds."""
@@ -3987,6 +4115,18 @@ class SphericalSurface3D(PeriodicalSurface):
 
                   ]
         return volmdlr.core.BoundingBox.from_points(points)
+
+    def get_circle_at_z(self, z_position: float):
+        """
+        Gets a circle on the sphere at given z position < radius.
+
+        :param z_position: circle's z position.
+        :return: circle 3D at given z position.
+        """
+        center1 = self.frame.origin.translation(self.frame.w * z_position)
+        circle_radius = math.sqrt(self.radius ** 2 - center1.point_distance(self.frame.origin) ** 2)
+        circle = curves.Circle3D(volmdlr.Frame3D(center1, self.frame.u, self.frame.v, self.frame.w),  circle_radius)
+        return circle
 
     def contour2d_to_3d(self, contour2d):
         """
@@ -4575,7 +4715,7 @@ class SphericalSurface3D(PeriodicalSurface):
             ax = fig.add_subplot(111, projection='3d')
 
         self.frame.plot(ax=ax, ratio=self.radius)
-        for circle in self._circle_generatrices(50):
+        for circle in self._circle_generatrices(50) + self._circle_generatrices_xy(50):
             circle.plot(ax, edge_style)
         return ax
 
@@ -4759,6 +4899,8 @@ class SphericalSurface3D(PeriodicalSurface):
         if circle_plane.point_distance(self.frame.origin) > self.radius:
             return []
         circle_plane_intersections = self.plane_intersections(circle_plane)
+        if circle_plane_intersections and isinstance(circle_plane_intersections[0], volmdlr.Point3D):
+            return []
         intersections = circle_plane_intersections[0].circle.circle_intersections(circle)
         return intersections
 
@@ -4816,6 +4958,42 @@ class SphericalSurface3D(PeriodicalSurface):
         :return: list containing the intersection points.
         """
         return self.ellipse_intersections(fullarcellipse.ellipse)
+
+    def _spherical_intersection_points(self, spherical_surface: 'SphericalSurface3D'):
+        """
+        Gets the points of intersections between the spherical surface and the toroidal surface.
+
+        :param spherical_surface: other Conical Surface 3d.
+        :return: points of intersections.
+        """
+        cyl_generatrices = self._circle_generatrices(200) + self._circle_generatrices_xy(200)
+        intersection_points = []
+        for gene in cyl_generatrices:
+            intersections = spherical_surface.edge_intersections(gene)
+            for intersection in intersections:
+                if not intersection.in_list(intersection_points):
+                    intersection_points.append(intersection)
+        return intersection_points
+
+    def sphericalsurface_intersections(self, spherical_surface: 'SphericalSurface3D'):
+        """
+        Cylinder Surface intersections with a Spherical surface.
+
+        :param spherical_surface: intersecting sphere.
+        :return: list of intersecting curves.
+        """
+        intersection_points = self._spherical_intersection_points(spherical_surface)
+        if not intersection_points:
+            return []
+        inters_points = vm_common_operations.separate_points_by_closeness(intersection_points)
+        curves_ = []
+        for list_points in inters_points:
+            bspline = edges.BSplineCurve3D.from_points_interpolation(list_points, 4, centripetal=False)
+            if isinstance(bspline.simplify, edges.FullArc3D):
+                curves_.append(bspline.simplify)
+                continue
+            curves_.append(bspline)
+        return curves_
 
 
 class RuledSurface3D(Surface3D):
@@ -5658,8 +5836,7 @@ class BSplineSurface3D(Surface3D):
     def __init__(self, degree_u: int, degree_v: int, control_points: List[volmdlr.Point3D], nb_u: int, nb_v: int,
                  u_multiplicities: List[int], v_multiplicities: List[int], u_knots: List[float], v_knots: List[float],
                  weights: List[float] = None, name: str = ''):
-        self.ctrlpts = npy.asarray([npy.asarray([*point], dtype=npy.float64) for point in control_points],
-                                   dtype=npy.float64)
+        self.ctrlpts = npy.asarray(control_points)
         self.degree_u = int(degree_u)
         self.degree_v = int(degree_v)
         self.nb_u = int(nb_u)
@@ -6569,7 +6746,7 @@ class BSplineSurface3D(Surface3D):
         if res.fun < 1e-6:
             return volmdlr.Point2D(*res.x)
 
-        point3d_array = npy.array([point3d[0], point3d[1], point3d[2]], dtype=npy.float64)
+        point3d_array = npy.array(point3d)
         delta_bound_x = max_bound_x - min_bound_x
         delta_bound_y = max_bound_y - min_bound_y
         x0s = [((min_bound_x + max_bound_x) / 2, (min_bound_y + max_bound_y) / 2),
@@ -6700,7 +6877,7 @@ class BSplineSurface3D(Surface3D):
         points = []
         for point in linesegment2d.discretization_points(number_points=20):
             point3d = self.point2d_to_3d(point)
-            if not volmdlr.core.point_in_list(point3d, points):
+            if not point3d.in_list(points):
                 points.append(point3d)
         if len(points) < 2:
             return None
@@ -6903,7 +7080,7 @@ class BSplineSurface3D(Surface3D):
         points = []
         for point in bspline_curve2d.discretization_points(number_points=number_points):
             point3d = self.point2d_to_3d(point)
-            if not volmdlr.core.point_in_list(point3d, points):
+            if not point3d.in_list(points):
                 points.append(point3d)
         if len(points) < bspline_curve2d.degree + 1:
             return None
@@ -6919,7 +7096,7 @@ class BSplineSurface3D(Surface3D):
         tol = 1e-6 if arc3d.length() > 1e-5 else 1e-8
         for point3d in arc3d.discretization_points(number_points=number_points):
             point2d = self.point3d_to_2d(point3d, tol)
-            if not volmdlr.core.point_in_list(point2d, points):
+            if not point2d.in_list(points):
                 points.append(point2d)
         start = points[0]
         end = points[-1]
@@ -7065,7 +7242,7 @@ class BSplineSurface3D(Surface3D):
         for point in self.control_points[1:]:
             vector = point - points[0]
             is_colinear = any(vector.is_colinear_to(other_vector) for other_vector in vector_list)
-            if not point_in_list(point, points) and not is_colinear:
+            if not point.in_list(points) and not is_colinear:
                 points.append(point)
                 vector_list.append(vector)
                 if len(points) == 3:
@@ -7785,7 +7962,7 @@ class BSplineSurface3D(Surface3D):
         :return: B-spline surface.
         :rtype: BSplineSurface3D
         """
-        points = npy.asarray([npy.asarray([*point], dtype=npy.float64) for point in points_3d], dtype=npy.float64)
+        points = npy.asarray(points_3d)
 
         ctrlpts, knots_u, knot_multiplicities_u, knots_v, knot_multiplicities_v = \
             interpolate_surface(points, size_u, size_v, degree_u, degree_v)
@@ -7836,7 +8013,7 @@ class BSplineSurface3D(Surface3D):
         # number of data points, s + 1 > number of control points, m + 1
         num_cpts_v = kwargs.get('ctrlpts_size_v', size_v - 1)
 
-        points = npy.asarray([npy.asarray([*point], dtype=npy.float64) for point in points_3d], dtype=npy.float64)
+        points = npy.asarray(points_3d)
 
         ctrlpts, knots_u, knot_multiplicities_u, knots_v, knot_multiplicities_v = \
             approximate_surface(points, size_u, size_v, degree_u, degree_v,